Refactor recognisers into a Python module
