+++ setup.py
@@ -1,10 +1,9 @@
 #!/usr/bin/env python
-
-from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
+from distutils.core import setup
 
 d = generate_distutils_setup(
-    install_requires=['dlib', 'tensorflow-gpu', 'numpy', 'opencv-python']
+    packages=['ros_people_model'],
+    package_dir={'': 'src'}
 )
-
 setup(**d)
