+++ src/ros_people_model/__init__.py
@@ -1,0 +1,1 @@
+from . import recogniser_emotion, recogniser_eye_state, recogniser_face_id, recogniser_landmarks, math
